puppet

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,6 @@
 #kill process killmenow
-exec { 'pkill':
-command => 'pkill killmenow',
-provider => 'shell',
+exec { 'kill_process':
+  command     => 'pkill killmenow',
+  refreshonly => true,
+  onlyif      => 'pgrep killmenow',
 }
